Release 0.7 (#154)
Workflow:
*u All Workflows - use Java 17

Build Updates:
*u gradle-wrapper.properties - upgrade Gradle 8.4
*u dependencies.gradle - upgrade Guava 32.1.3, add findbugs artifact coordinate
*u artifacts.gradle - add artifact version definitions for new modules, update artifact versions to 0.7
*u settings.gradle - include new modules, upgrade Dokka 1.9.0
*c Arrays build.gradle
*c Arrays-ktx build.gradle
*c Format build.gradle
*c Pairs build.gradle
*u Lists-ktx build.gradle - add dependency on Arrays-ktx
*u Numbers build.gradle - replace Guava with FindBugs, add dependency on Pairs, replace test dependency on Lists with Arrays
*u Numbers build.gradle - reduce Branch coverage from 96% to 95.7%

Format:
* Move Format Package from Numbers Module to Format Module

Format Tests:
* Move Format Package from Numbers Module to Format Module

Pairs:
* Move Package Structs from Numbers Module to Pairs Module

Pairs Tests:
* Move Package Structs from Numbers Module to Pairs Module
* Move TimeMeasurement to Primes Test Package

Arrays:
* Move Arrays Package from Lists Module to Arrays Module
* Move Arrays Package from Lists-ktx Module to Arrays-ktx Module
* Move BigIntSumBuffer to Arrays Module

Arrays Tests:
* Move Arrays Package from Lists Module to Arrays Module
* Move Arrays Package from Lists-ktx Module to Arrays-ktx Module
* Move BigIntSumBuffer to Arrays Module

Documentation:
*u README.md

Lists-ktx:
*u Usages of BigIntSumBuffer

Tests:
*u LongArrayExtTest - add test cases

Workflow:
*u coverage_report.yml - update path to include all modules
*u publish_release.yml - update path to include all modules

Numbers Primes:
*u ShortPrimeCache - remove unreachable branches
*u ShortPrimeCacheTest - add test case

+++ lists-ktx/src/main/kotlin/mathtools/lists/ByteList.kt
@@ -1,5 +1,6 @@
 package mathtools.lists
 
+import mathtools.arrays.sum.BigIntSumBuffer
 import java.math.BigInteger
 
 /** Functions of Byte typed Lists
@@ -139,7 +140,7 @@ object ByteList {
 			else -> BigInteger.ZERO
 		}
 		val buffer = BigIntSumBuffer()
-		for (i in list) buffer.add(i.toLong())
+		for (i in list) buffer.add(i.toInt())
 		return buffer.sum
 	}
